Create get_yearly_token_scores.py
The purpose of this code is to gather all the frequencies of two tokens with differing spelling ('public' and 'catholic') across two data sets - all editions, and only one editions per work, in the ECCO corpus. I work with two additional files;  a previous project done by other students which has a list of all the works in ECCO (with some additional data, requires cleaning) and a list of ID conversions done by one member of COMHIS (previous project file includes the IDs in a different format to that of ECCO). Once the six tokens are gathered, I arrange them by yearly aggregates to create two files which then can be manually examined for trends.